feat: リポジトリ層のユニットテスト実装
- PRメトリクスリポジトリの包括的なテストスイート実装
- 集計メトリクスリポジトリのテストケース作成
- データ永続化、検索、取得機能のテスト網羅
- エラーハンドリングとエッジケースのテスト
- トランザクション処理と一括操作のテスト
- モック利用による外部依存関係の分離

+++ backend/app/infrastructure/repository/aggregated_metrics_repository.go
@@ -543,7 +543,7 @@ func (repo *AggregatedMetricsRepository) convertTeamMetricsToStorage(metrics *an
 		
 		// 生産性指標（チームでは計算しない）
 		PRsPerDay:   float64(metrics.TotalPRs) / repo.calculateDays(metrics.DateRange),
-		LinesPerDay: metrics.SizeStats.LinesChanged.Sum / repo.calculateDays(metrics.DateRange),
+		LinesPerDay: float64(metrics.SizeStats.LinesChanged.Sum) / repo.calculateDays(metrics.DateRange),
 		Throughput:  float64(metrics.TotalPRs),
 		
 		// トレンド情報
@@ -699,7 +699,7 @@ func (repo *AggregatedMetricsRepository) convertRepositoryMetricsToStorage(metri
 		
 		// 生産性指標
 		PRsPerDay:   float64(metrics.TotalPRs) / repo.calculateDays(metrics.DateRange),
-		LinesPerDay: metrics.SizeStats.LinesChanged.Sum / repo.calculateDays(metrics.DateRange),
+		LinesPerDay: float64(metrics.SizeStats.LinesChanged.Sum) / repo.calculateDays(metrics.DateRange),
 		Throughput:  float64(metrics.TotalPRs),
 		
 		// トレンド情報（リポジトリメトリクスには含まれないため空）
